server/requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.13
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-environ==0.4.4
django-filter==1.1.0
django-import-export==1.2.0
django-uuidfield==0.5.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
fitbit==0.3.0
geographiclib==1.49
geopy==1.16.0
gunicorn==19.8.1
honcho==1.0.1
idna==2.7
jdcal==1.4
kombu==4.2.1
kwargs-only==1.0.0
Markdown==2.6.11
numpy==1.15.2
oauthlib==2.1.0
odfpy==1.4.0
openpyxl==2.5.14
psycopg2-binary==2.7.4
pyfcm==1.4.5
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
Shapely==1.6.4.post2
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
timezonefinder==3.1.0
unicodecsv==0.14.1
urllib3==1.23
vine==1.1.4
xlrd==1.2.0
xlwt==1.3.0
